requirements.txt
flask>=2.0.0
werkzeug>=2.0.3
torch>=2.0.0
torchvision>=0.15.0
Pillow>=8.0.0
numpy>=1.21.0
matplotlib>=3.4.0
pandas>=1.3.0
scikit-learn>=0.24.0
tqdm>=4.62.0
streamlit>=1.32.0 